+++ src/ClassBuilder.js
@@ -24,6 +24,8 @@
  WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  ------------------------------------------------------------------------------
 */
+'use strict';
+
 (function(g, undefined) {
 
     // Register JOII 'namespace'.
@@ -75,10 +77,12 @@
             // types are treated statically throughout all instances.
             scope_in = g.JOII.Compat.extend(true, {}, scope_in);
 
-            g.JOII.CreateProperty(scope_in, '__joii__', (this.__joii__));
-            g.JOII.CreateProperty(scope_out, '__joii__', (this.__joii__));
+            if (typeof this !== 'undefined') {
+                g.JOII.CreateProperty(scope_in, '__joii__', (this.__joii__));
+                g.JOII.CreateProperty(scope_out, '__joii__', (this.__joii__));
+            }
 
-            if (typeof(this.__joii__) === 'object') {
+            if (typeof this !== 'undefined' && typeof(this.__joii__) === 'object') {
 
                 // Can we be instantiated?
                 if (this.__joii__.is_abstract === true) {
@@ -106,12 +110,17 @@
             // we've been executed like a function rather than being instantiated.
             if (typeof(this) === 'undefined' || typeof(this.__joii__) === 'undefined') {
                 // If the method __call exists, execute it and return its result.
-                if (typeof(body.__call) === 'function') {
-                    var result = body.__call.apply(body, arguments);
-                    if (result === body) {
-                        throw '__call cannot return itself.';
+
+                for (var c in g.JOII.Config.callables) {
+                    if (g.JOII.Config.callables.hasOwnProperty(c)) {
+                        if (typeof(body[g.JOII.Config.callables[c]]) === 'function') {
+                            var result = body[g.JOII.Config.callables[c]].apply(body, arguments);
+                            if (result === body) {
+                                throw g.JOII.Config.callables[c] + ' cannot return itself.';
+                            }
+                            return result;
+                        }
                     }
-                    return result;
                 }
                 throw 'This class cannot be called as a function because it\'s lacking the __call method.';
             }
@@ -120,8 +129,14 @@
             scope_in.__api__ = scope_out;
 
             // Does the class defintion have a constructor? If so, run it.
-            if (typeof(scope_in.__construct) === 'function') {
-                scope_in.__construct.apply(scope_in, arguments);
+            for (var c in g.JOII.Config.constructors) {
+                if (g.JOII.Config.constructors.hasOwnProperty(c)) {
+                    var cc = g.JOII.Config.constructors[c];
+                    if (typeof(scope_in[cc]) === 'function') {
+                        scope_in[cc].apply(scope_in, arguments);
+                        break;
+                    }
+                }
             }
 
             // Are we attempting to instantiate an abstract class?